backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.1
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
python-dateutil==2.8.2
